Move license availability table to its own component
